lib: Add math.rem
math.rem computes the remainder of a truncated division, in which the
sign of the remainder is that of the dividend.  It is the most popular
implementation of the % operator, according to Wikipedia [1].

I thought about writing a math.mod as well, but it seems that there are
several competing alternatives to that.  Knuth's floored division seems
popular, as well as Euclidean division.  So I'll add it when there is
actual need for it.

[1]: https://en.wikipedia.org/wiki/Modulo_operation#In_programming_languages