ci: Configure nixos-20.09 channel explicitly
Ensure that we're using the latest revision of the channel, instead of
whatever revision the OS image was built with.  I hope it's reasonable
to expect users to be up to date.

If we end up rolling this back, we also need to rollback the previous
commit, as the sourcehut image comes with Cachix 0.5.1 which doesn't
have `watch-exec`.